% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{big4year}
\alias{big4year}
\title{Data on the big four EPL Teams}
\format{A data frame with 21 rows and 4 columns}
\description{
The English Premier League (EPL) is a famous professional league for association
soccer clubs in the UK. The so-called "Big Four" soccer clubs are Arsenal, Chelsea,
Liverpool, and Manchester United, have dominated the top four spots since the 1996-1997
season. Wikipedia documented the results of the "Big Four" since the start of the
Premier League in the 1992-93 season. The rankings of these four English Premier
League teams from the 1992-93 season to the2012-13 season are listed in this data frame.
The first row of the data means that there is one season that Arsenal ranked at the
top of EPL, Chelsea the second, Manchester United the third and Liverpool the fourth.
We may test the hypothesis that the rankings observed are random. On the other hand,
through the use of modeling we may try to determine how the rankings cluster.
}
\details{
\itemize{
  \item The ranks of the "Big Four"(Arsenal, Chelsea, Liverpool, and Manchester United) clubs

}
}
\keyword{datasets}

