% Generated by roxygen2 (4.0.2): do not edit by hand
\name{case2freq}
\alias{case2freq}
\title{Convert raw ranking data(case form) to ranking data with rank frequencies}
\usage{
case2freq(x)
}
\arguments{
\item{x}{A data frame or a matrix(case form), each row is a rank.}
}
\value{
A data frame, each row contains a rank and the corresponding frequency.
}
\description{
Convert raw ranking data(case form) to ranking data with rank frequencies
}
\examples{
data(APA)
cases = freq2case(APA, freq.col = 1)
freqs = case2freq(cases)
}
\author{
Li Qinglong <liqinglong0830@163.com>
}

