% Generated by roxygen2 (4.0.2): do not edit by hand
\name{compatible.rankings}
\alias{compatible.rankings}
\title{Set of compatible rankings}
\usage{
compatible.rankings(input_ranking)
}
\arguments{
\item{input_ranking}{a vector, using NAs to stand for the missing ranks}
}
\value{
a matrix or a vector, each column is a complete ranking.
}
\description{
This function generates the set of complete permutations compatible
with the input_ranking.
}
\examples{
u_star = c(2, NA, 3, 4, 1)
C_set = compatible.rankings(u_star)
is.compatible(C_set, u_star)
}
\author{
Li Qinglong <liqinglong0830@163.com>
}

