% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{Croon}
\alias{Croon}
\title{Croon data on goals of the Government}
\format{A matrix with 24 rows and 5 columns}
\description{
M.A. Croon conducted a survey in Germany on the political goals on the Government.
The data set contains 2262 rankings, in the aggregated form, of four goals.
}
\details{
\itemize{
  \item 1, Maintain order in nation
  \item 2, Give people more say in Government decisions
  \item 3, Fight rising prices
  \item 4, Protect freedom of speech
  each column is a rank of the four items and the frequency of the rank
}
}
\keyword{datasets}

