% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generate.combs}
\alias{generate.combs}
\title{Generate all possible combinations of k elements out of n}
\usage{
generate.combs(n, k = 1)
}
\arguments{
\item{n}{number of the whole elements}

\item{k}{number of elements to choose (default 1)}
}
\value{
a matrix or a vector
}
\description{
This function generates all combinations of n elements taken k at a time. It is just
a reference to the combn function in utils package.
}
\examples{
generate.combs(10, 6)
}
\author{
Scott Chasalow wrote the original in 1994 for S;
R package combinat and documentation by Vince Carey stvjc@channing.harvard.edu;
small changes by the R core team.
}
\seealso{
utils::combn
}

