% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generate.perms}
\alias{generate.perms}
\title{Generate all possible permutations of k elements out of n}
\usage{
generate.perms(n, k = n, vec = 1:n)
}
\arguments{
\item{n}{number of the whole elements}

\item{k}{number of elements to permute (default the same as n)}

\item{vec}{the source vector of length n (default as c(1:n))}
}
\value{
a matrix or a vector, each row is a permutation
}
\description{
This function generates all permutations of n elements taken k at a time.
An efficient way to generate n! permutations of vec
in a Lexicographic order without recursive algorithm
}
\examples{
generate.perms(10, 6)
}
\author{
Li Qinglong <liqinglong0830@163.com>
}
\seealso{
generate.combn
}

