% Generated by roxygen2 (4.0.2): do not edit by hand
\name{incomplete.rankings}
\alias{incomplete.rankings}
\title{Generate all incomplete rankings of k elements out of n complete elements}
\usage{
incomplete.rankings(n, k)
}
\arguments{
\item{n}{number of complete elements}

\item{k}{number of non-missing elements in the incomplete ranking}
}
\value{
a matrix each of whose column is a possible incomplete ranking
}
\description{
This function generates all the possible incomplete rankings of
k elements out of n complete elements. E.g. c(1, 2, NA) is an
incomplete ranking out with 2 non-missing elements of 3 elements
}
\examples{
incomplete.rankings(5, 3)
}
\author{
Li Qinglong <liqinglong0830@163.com>
}

