% Generated by roxygen2 (4.0.2): do not edit by hand
\name{is.compatible}
\alias{is.compatible}
\title{Is compatible with the candidate_ranking}
\usage{
is.compatible(complete_ranking, candidate_ranking)
}
\arguments{
\item{complete_ranking}{a matrix or a vector, each column is a complete ranking.}

\item{candidate_ranking}{a vector, using NA to stand for the missing ranks.}
}
\value{
a vector of TRUEs and FALSEs
}
\description{
This function judges whether the complete rangkings are compatible with
an incomplete ranking.
}
\examples{
u_star = c(2, NA, 3, 4, 1)
C_set = compatible.rankings(u_star)
is.compatible(C_set, u_star)
}
\author{
Li Qinglong <liqinglong0830@163.com>
}

