% Generated by roxygen2 (4.0.2): do not edit by hand
\name{msci}
\alias{msci}
\title{Mean, Standard deviation, Confident Intervals}
\usage{
msci(df, sig = 0.05)
}
\arguments{
\item{df}{a dafa frame, each column is a sampling trace.}

\item{sig}{significance level, defaul is 0.05}
}
\value{
a data frame with MSCI of each trace
}
\description{
Get the Mean, Standard deviation and Confidence Intervals of a sampling trace
}
\author{
Li Qinglong <liqinglong0830@163.com>
}

