% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotDecROC.R
\name{plotDecROC}
\alias{plotDecROC}
\title{ROC Curves Plot of Decision Tree}
\usage{
plotDecROC(mytree)
}
\arguments{
\item{mytree}{a list returned by the decTreexxx function contains the information of the decision tree.}
}
\description{
This function will draw ROC curves based on the tree information list.
}
\author{
Li Qinglong <liqinglong0830@163.com>
}
\seealso{
plotDecTree
}

