% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotDecTree.R
\name{plotDecTree}
\alias{plotDecTree}
\title{Decision Tree Plot}
\usage{
plotDecTree(mytree, width = 1000, height = 750, uniform = TRUE,
  branch = 0, margin = 0.1, nspace = -1, minbranch = 0.3, all = FALSE,
  fancy = TRUE, use.n = TRUE, fwidth = 0.5, fheight = 0.5)
}
\arguments{
\item{mytree}{a list returned by the decTreexxx function contains the information of the decision tree.}

\item{width,}{etc. parameters for ploting, see rpart package for more details. You can try serveral times
for the best display.}

\item{height}{height}

\item{uniform}{TRUE or FALSE}

\item{branch}{branch}

\item{margin}{margin}

\item{nspace}{nspace}

\item{minbranch}{minbranch}

\item{all}{all}

\item{fancy}{fancy}

\item{use.n}{use.n}

\item{fwidth}{fwidth}

\item{fheight}{fheight}
}
\value{
the plot will be save as a png file in your working directory.
}
\description{
This function will draw decision tree based on the tree information list. All the plotting functions
come frome rpart package.
}
\author{
Li Qinglong <liqinglong0830@163.com>
}
\seealso{
plotDecROC
}

