% Generated by roxygen2 (4.0.2): do not edit by hand
\name{repmat}
\alias{repmat}
\title{Replicate and tile array}
\usage{
repmat(A, M, N)
}
\arguments{
\item{A}{Matrix or vector to repeat.}

\item{M}{Number of row repititions.}

\item{N}{Number of column repititions.}
}
\value{
a matrix of M-by-N tiling of A.
}
\description{
Just the same usage as repmat function in Matlab. This kind of functions
can be easily found on the web.
}
\examples{
repmat(c(1, 2), 6, 8)
}

