% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{Scores}
\alias{Scores}
\title{Test scores in Language and Arithmetic for a group of 9 students}
\format{A data frame of 2 variables}
\description{
See Page 28, Lehmann, E. (1975). Nonparametrics: Statistical Methods based on Ranks. McGrawHill, New York\cr
}
\details{
\itemize{
  \item Lang.    Language scores of 9 students.
  \item Arith.   Arithmetic scores of 9 students.

}
}
\keyword{datasets}

