\name{StatMethRank-package}
\alias{StatMethRank-package}
\alias{StatMethRank}
\docType{package}
\title{
Statistical Methods for Ranking Data
}
\description{
Statistical Methods for Ranking Data
}
\details{
\tabular{ll}{
Package: \tab StatMethRank\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-05-28\cr
License: \tab GPL-2\cr
}
Functions and datasets to support Alvo and Yu (2014), 'Statistical Methods for Ranking Data'
}
\author{
Li Qinglong

Maintainer: Li Qinglong <liqinglong0830@163.com>
}
\references{
Statistical Methods for Ranking Data, Alvo and Yu (2014), (\url{http://web.hku.hk/~plhyu/StatMethRank/})
}

\keyword{ ranking }
\seealso{
pmr   (\url{http://cran.r-project.org/web/packages/pmr/index.html})
}
